:recycle: refactor `callbacks` module
