+++ fastapi_async_langchain/callbacks/llm.py
@@ -1,0 +1,19 @@
+from typing import Any
+
+from .base import AsyncStreamingResponseCallback, AsyncWebsocketCallback
+
+
+class AsyncLLMChainStreamingCallback(AsyncStreamingResponseCallback):
+    """AsyncStreamingResponseCallback handler for LLMChain."""
+
+    async def on_llm_new_token(self, token: str, **kwargs: Any) -> None:
+        """Run on new LLM token. Only available when streaming is enabled."""
+        await self.send(token)
+
+
+class AsyncLLMChainWebsocketCallback(AsyncWebsocketCallback):
+    """AsyncWebsocketCallback handler for LLMChain."""
+
+    async def on_llm_new_token(self, token: str, **kwargs: Any) -> None:
+        """Run on new LLM token. Only available when streaming is enabled."""
+        await self.websocket.send_json({**self.response.dict(), **{"message": token}})
